HBX-1660: Move class 'org.hibernate.tool.hbm2x.pojo.SkipBackRefPropertyIterator' to 'org.hibernate.tool.internal.util.SkipBackRefPropertyIterator'
