+++ orm/src/main/java/org/hibernate/tool/internal/util/SkipBackRefPropertyIterator.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.hbm2x.pojo;
+package org.hibernate.tool.internal.util;
 
 import java.util.Iterator;
 import java.util.NoSuchElementException;
